requirements.txt
numpy==1.21.5
Pillow==9.4.0
scikit_learn==1.2.1
scipy==1.8.0
six==1.16.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.60.0
